Revamp java wrapper by changing class path and Bump connecter version to 0.1.4
