+++ gradle.properties
@@ -1,4 +1,4 @@
 org.gradle.caching=true
-group=io.ballerinax.webhook
+group=org.ballerinalang.googleapis.drive
 version=0.1.4-SNAPSHOT
 ballerinaLangVersion=2.0.0-beta.1-20210531-213400-7cdb0fb5
